feat: forenkle bruken av tooltip etter label i text-input
Bruker InputGroup internt for å lette på plasseringen av tooltip.

BREAKING CHANGE:
- `variant` er fjernet. Bruk `labelProps={{ variant }}` i stedet.
- `BaseInputField` er erstattet av `BaseTextInput`.
- Du må importere CSS for
[input-group](https://github.com/fremtind/jokul/tree/main/packages/input-group).
