+++ packages/text-input-react/README.md
@@ -13,14 +13,19 @@ Se portalen for [bruk og prinsipper](https://jokul.fremtind.no/komponenter/texti
 import { TextInput, TextAera } from "@fremtind/jkl-text-input-react";
 
 // Importer stilark via JavaScript med CSS-loader.
+import "@fremtind/jkl-input-group/input-group.min.css";
 import "@fremtind/jkl-text-input/text-input.min.css";
 
-// For handlingsknapper må du også ha denne CSSen importert.
+// For enkelte features må du også ha denne CSSen importert.
 import "@fremtind/jkl-icon-button/icon-button.min.css";
+import "@fremtind/jkl-tooltip/tooltip.min.css";
 ```
 
 ```scss
 // Eller importer stilark via SCSS.
+@use "@fremtind/jkl-input-group/input-group";
 @use "@fremtind/jkl-text-input/text-input";
+
 @use "@fremtind/jkl-icon-button/icon-button";
+@use "@fremtind/jkl-tooltip/tooltip";
 ```
