+++ packages/text-input-react/src/BaseTextArea.tsx
@@ -1,0 +1,144 @@
+import React, {
+    type FocusEvent,
+    forwardRef,
+    type RefObject,
+    type TextareaHTMLAttributes,
+    useRef,
+    useState,
+    useEffect,
+} from "react";
+
+type Counter = {
+    /** Antall tegn før telleren når maksimum og vi viser en feilmelding */
+    maxLength: number;
+    /**
+     * Med teller vises en progress-bar i bunnen av tekstfeltet som krymper
+     * ned fra 100% (null tegn skrevet) til 0% (maks antall tegn skrevet).
+     * Om du vil skjule den, sett denne propen til true.
+     *
+     * @default false
+     */
+    hideProgress?: boolean;
+};
+
+export interface BaseTextAreaProps extends Omit<TextareaHTMLAttributes<HTMLTextAreaElement>, "children"> {
+    /**
+     * Viser en teller i tekstfeltet når det har fokus.
+     * Denne er ulik den innebyggede maxLength i at den ikke setter en
+     * stopper for å gå over grensen, slik at brukeren kan fullføre en
+     * tankerekke.
+     */
+    counter?: Counter;
+    /** @deprecated Foretrekk counter-propen sin maxLength for å la brukerne fullføre en tankerekke før de redigerer seg innenfor maksgrensen */
+    maxLength?: number;
+    /** Sett antall rader skjemafeltet ekspanderes til ved focus. Innholdet scroller om feltet fylles med mer innhold enn det er plass til. */
+    rows?: number;
+    inline?: boolean;
+    startOpen?: boolean;
+    /** Gjør så skjemafeltet ekspanderer seg til å vise alt innhold i stedet for å scrolle */
+    autoExpand?: boolean;
+}
+
+export const BaseTextArea = forwardRef<HTMLTextAreaElement, BaseTextAreaProps>((props, ref) => {
+    const {
+        autoExpand,
+        counter,
+        onBlur,
+        onFocus,
+        rows,
+        startOpen,
+        style,
+        value,
+        "aria-invalid": ariaInvalid,
+        ...rest
+    } = props;
+
+    const [textAreaFocused, setTextAreaFocused] = useState(false);
+    const internalRef = useRef<HTMLTextAreaElement>(null);
+    const textAreaRef = (ref as RefObject<HTMLTextAreaElement>) || internalRef;
+
+    useEffect(() => {
+        const textAreaElement = textAreaRef.current;
+        if (textAreaElement) {
+            if (!autoExpand) {
+                textAreaElement.style.height = "";
+                return;
+            }
+
+            if (textAreaFocused || value) {
+                textAreaElement.style.height = "auto"; // Sett til auto før scrollhøyden leses, sånn at redusering av høyde ved sletting av tekst fungerer
+                textAreaElement.style.height = `${textAreaElement.scrollHeight}px`;
+            } else {
+                textAreaElement.style.height = "";
+            }
+        }
+    }, [autoExpand, textAreaRef, value, textAreaFocused]);
+
+    function handleOnFocus(e: FocusEvent<HTMLTextAreaElement>) {
+        setTextAreaFocused(true);
+        if (onFocus) {
+            onFocus(e);
+        }
+    }
+
+    function handleOnBlur(e: FocusEvent<HTMLTextAreaElement>) {
+        setTextAreaFocused(false);
+        if (onBlur) {
+            onBlur(e);
+        }
+    }
+
+    const counterCurrent: number = textAreaRef.current?.value.length || 0;
+    const counterTotal: number = counter?.maxLength || 0;
+    const progressCurrent: number = counterTotal - counterCurrent;
+    function calculatePercentage(current: number, total: number): number {
+        if (current <= 0) {
+            return 0;
+        }
+        return total === 0 ? 0 : (current * 100) / total;
+    }
+    const counterLabel =
+        counter && counterCurrent > counterTotal
+            ? `Du har skrevet ${counterCurrent - counterTotal} tegn for mye`
+            : undefined;
+
+    const invalid = Boolean(ariaInvalid || counterLabel);
+
+    const overflowStyle = {
+        overflowX: autoExpand ? "hidden" : undefined, // Must set overflowX hidden for Firefox https://stackoverflow.com/a/22700700
+    } as React.CSSProperties;
+
+    return (
+        <div className="jkl-text-input-wrapper" data-invalid={invalid} data-has-content={counterCurrent > 0}>
+            <textarea
+                aria-invalid={invalid}
+                className={`jkl-text-area__text-area jkl-text-area__text-area--${rows}-rows`}
+                onBlur={handleOnBlur}
+                onFocus={handleOnFocus}
+                ref={textAreaRef}
+                style={{ ...style, ...overflowStyle }}
+                value={value}
+                {...rest}
+            />
+            {counter && (
+                <div className="jkl-text-area__counter" aria-hidden="true">
+                    <div className="jkl-text-area__counter-count">
+                        {counterCurrent}&nbsp;/&nbsp;{counterTotal}
+                    </div>
+                    {!counter.hideProgress && (
+                        <div
+                            className="jkl-text-area__counter-progress"
+                            style={{
+                                ["--progress-width" as string]: `${calculatePercentage(
+                                    progressCurrent,
+                                    counterTotal,
+                                )}%`,
+                            }}
+                        />
+                    )}
+                </div>
+            )}
+        </div>
+    );
+});
+BaseTextArea.displayName = "BaseTextArea";
